# contracts/backend-api/1.0.0/openapi.json
{
  "components": {
    "responses": {
      "ContractError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "使用稳定错误代码表示的失败。"
      }
    },
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "fieldErrors": {
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "schemaVersion": {
            "const": "1.0.0"
          }
        },
        "required": [
          "schemaVersion",
          "error"
        ],
        "type": "object"
      },
      "ApiFieldError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "DocumentWrite": {
        "additionalProperties": false,
        "properties": {
          "assetIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "title": "Assetids",
            "type": "array"
          },
          "baseRevision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baserevision"
          },
          "contractFamily": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Contractfamily",
            "type": "string"
          },
          "contractVersion": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Contractversion",
            "type": "string"
          },
          "documentKind": {
            "enum": [
              "creator-workspace",
              "gm-tabletop-document",
              "character-save"
            ],
            "title": "Documentkind",
            "type": "string"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "mutationId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Mutationid",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "mutationId",
          "documentKind",
          "contractFamily",
          "contractVersion",
          "baseRevision",
          "assetIds",
          "payload"
        ],
        "title": "DocumentWrite",
        "type": "object"
      },
...
      "LifecycleWrite": {
        "additionalProperties": false,
        "properties": {
          "baseRevision": {
            "minimum": 1.0,
            "title": "Baserevision",
            "type": "integer"
          },
          "mutationId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Mutationid",
            "type": "string"
          }
        },
        "required": [
          "mutationId",
          "baseRevision"
        ],
        "title": "LifecycleWrite",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "siteSession": {
        "in": "header",
        "name": "X-PbDH-Session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "PbDH Platform 的身份、云文档、托管媒体和资源市场 HTTP 合约。领域 payload 仍由各自 Platform Contract 管理。",
    "title": "PbDH Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cloud/documents": {
      "get": {
        "operationId": "list_documents",
        "parameters": [
          {
            "in": "query",
            "name": "documentKind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "creator-workspace",
                    "gm-tabletop-document",
                    "character-save"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Documentkind"
            }
          },
          {
            "in": "query",
            "name": "includeDeleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Includedeleted",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Documents",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "cloud-documents"
        ]
      }
    },
    "/api/cloud/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseRevision",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Baserevision",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "cloud-documents"
        ]
      },
      "get": {
        "operationId": "get_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Document",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "cloud-documents"
        ]
      },
      "put": {
        "operationId": "put_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Document",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Put Document",
        "tags": [
          "cloud-documents"
        ]
      }
    },
    "/api/cloud/documents/{document_id}/media/{asset_id}": {
      "get": {
        "operationId": "get_document_media",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "WebP 图片字节。"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Get Document Media",
        "tags": [
          "cloud-documents"
        ]
      }
    },
    "/api/cloud/documents/{document_id}/restore": {
      "post": {
        "operationId": "restore_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Restore Document",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Restore Document",
        "tags": [
          "cloud-documents"
        ]
      }
    },
    "/api/cloud/documents/{document_id}/trash": {
      "post": {
        "operationId": "trash_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trash Document",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Trash Document",
        "tags": [
          "cloud-documents"
        ]
      }
    },
    "/api/cloud/media/{asset_id}": {
      "put": {
        "operationId": "prepare_media",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "image/webp": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Prepare Media",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Prepare Media",
        "tags": [
          "cloud-documents"
        ]
      }
    },
...
  },
...
}